requirements.txt
python-dotenv>=1.0.1
streamlit>=1.33.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.35.0
faiss-cpu>=1.7.4
reportlab>=4.0.9

